feat(kubricate): add GenerateCommand for generating stacks and update CLI command
