+++ packages/kubricate/src/load-config.ts
@@ -1,0 +1,27 @@
+import { loadConfig } from "unconfig";
+import { MARK_CHECK } from "./constant.js";
+import { KubricateConfig } from "./config.js";
+import c from 'ansis';
+
+export interface LoadConfigOptions {
+  root: string;
+  config?: string;
+  outDir: string;
+}
+
+export async function getConfig(options: LoadConfigOptions): Promise<KubricateConfig> {
+  const result = await loadConfig<KubricateConfig>({
+    cwd: options.root,
+    sources: [
+      {
+        files: options.config || 'kubricate.config',
+        // Allow all JS/TS file extensions except JSON
+        extensions: ['mts', 'cts', 'ts', 'mjs', 'cjs', 'js'],
+      },
+    ],
+    merge: false,
+  });
+  if (result.sources.length)
+    console.log(c.green`${MARK_CHECK} Config loaded from ${result.sources.join(', ')}`);
+  return result.config;
+}
